convert viewTable component to tsx and use better typing
