+++ src/publisher/views/components/DimensionPreviewTable.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Table from '../../../shared/views/components/Table';
-import { CSVHeader, ViewDTO } from '../../../shared/dtos/view-dto';
+import { ColumnHeader, ViewDTO } from '../../../shared/dtos/view-dto';
 import { useLocals } from '../../../shared/views/context/Locals';
 import T from '../../../shared/views/components/T';
 import { SingleLanguageDimension } from '../../../shared/dtos/single-language/dimension';
@@ -8,7 +8,7 @@ import { SourceType } from '../../../shared/enums/source-type';
 
 export type DimensionPreviewTableProps = {
   headers: Array<
-    CSVHeader & {
+    ColumnHeader & {
       source_type: SourceType & 'line_number';
     }
   >;
